+++ cmd/prune.go
@@ -57,15 +57,9 @@ func runPrune(cmd *cobra.Command, args []string) error {
 
 	ctx := context.Background()
 
-	stack, err := config.Load(flagFile)
-	if err != nil {
-		return err
-	}
-	if flagProject != "" {
-		stack.Project = flagProject
-	}
+	stack, _ := config.Load(flagFile)
 
-	r, err := runtimeFrom(stack)
+	r, err := runtimeFromOptional(stack)
 	if err != nil {
 		return err
 	}